pyproject.toml
[project]
name = "marketing-graphrag"
version = "0.1.0"
description = "GraphRAG system for marketing agency analytics"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "llama-index>=0.10.0",
    "llama-index-graph-stores-neo4j>=0.2.0",
    "llama-index-llms-anthropic>=0.1.0",
    "llama-index-embeddings-voyageai>=0.1.0",
    "neo4j>=5.0.0",
    "anthropic>=0.18.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "streamlit>=1.31.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "tenacity>=8.0.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "reportlab>=4.0.0",
    "openpyxl>=3.1.0",
    "google-ads>=23.0.0",
    "facebook-business>=19.0.0",
    "apscheduler>=3.10.0",
    "slack-sdk>=3.0.0",
    "sendgrid>=6.0.0",
    "plotly>=5.18.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "config"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
